+++ history/service.go
@@ -6,6 +6,6 @@ import "github.com/Benyam-S/onepay/entity"
 type IService interface {
 	AddHistory(newOPHistory *entity.UserHistory) error
 	SearchHistories(key, orderBy string, methods []string, pageNum int64, columns ...string) []*entity.UserHistory
-	FindHistory(identifier int) (*entity.UserHistory, error)
+	FindHistory(identifier int64) (*entity.UserHistory, error)
 	AllUserHistories(userID string) []*entity.UserHistory
 }
